+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.5.0] - 2025-09-05
+### Added
+- Support for GET /users/{userId}/plans
+- Support for GET /2.0/users
+- Support for DELETE /users/{userId}/plans/{planId}
+
 ## [4.3.0] - 2025-26-27
 ### Added
 - ESLint Rule for enforcing type imports.